refactor(app-structure): organize app directory structure
- Reorganized app directory structure to improve maintainability and clarity.
