+++ src/components/articles/ArticleCard.tsx
@@ -1,8 +1,8 @@
 import Image from "next/image";
 
-import ArticleDetails from "@/components/cards/ArticleDetails";
+import ArticleDetails from "@/components/articles/ArticleDetails";
 import ReadMoreLink from "@/components/ui/ReadMoreLink";
-import { Article } from "@/lib/types";
+import { Article } from "@/types";
 
 const ArticleCard = (article: Article) => {
   const { imageUrl, title } = article;
