--- src/components/buttons/LoadMoreArticles.tsx
@@ -1,11 +1,0 @@
-const LoadMoreArticles = () => {
-  return (
-    <div className="mt-8 flex justify-center">
-      <button className="rounded-md bg-gray-200 px-6 py-2 text-gray-700 transition duration-300 hover:bg-gray-300">
-        Load More
-      </button>
-    </div>
-  );
-};
-
-export default LoadMoreArticles;
