+++ src/components/ui/Checkbox.tsx
@@ -1,4 +1,4 @@
-import React, { PropsWithChildren } from "react";
+import { PropsWithChildren } from "react";
 
 const Checkbox = ({ children }: PropsWithChildren) => {
   return (
