+++ src/components/ui/ReadMoreLink.tsx
@@ -2,8 +2,8 @@
 
 import Link from "next/link";
 
-import { useArticle } from "@/contexts/ArticleContext";
-import { Article } from "@/lib/types";
+import useArticle from "@/hooks/useArticle";
+import { Article } from "@/types";
 
 const ReadMoreLink = ({ article }: { article: Article }) => {
   const { setSelectedArticle } = useArticle();
