+++ src/types/newYorkTimesApi.d.ts
