+++ src/types/newsApi.d.ts
